Move work related stuff out of ElasticsearchFixtures
