+++ .buildkite/pipeline.yml
@@ -1,7 +1,4 @@
 steps:
-  - command: .buildkite/scripts/run_autoformat.py
-    label: "autoformat"
-  - wait
   - command: .buildkite/scripts/run_job.py internal_model
     label: "internal model"
   - command: .buildkite/scripts/run_job.py display
@@ -71,6 +68,9 @@ steps:
     label: "palette inferrer"
   - command: .buildkite/scripts/run_job.py snapshot_generator
     label: "snapshot generator"
+  - wait
+  - command: .buildkite/scripts/run_autoformat.py
+    label: "autoformat"
   - wait
   - label: cut release
     if: build.branch == "master"
