+++ common/elasticsearch/src/test/scala/uk/ac/wellcome/elasticsearch/test/fixtures/ElasticsearchFixtures.scala
@@ -1,31 +1,28 @@
 package uk.ac.wellcome.elasticsearch.test.fixtures
 
-import scala.concurrent.ExecutionContext.Implicits.global
-import scala.concurrent.Future
-import org.scalatest.concurrent.PatienceConfiguration.Timeout
-import org.scalatest.concurrent.{Eventually, IntegrationPatience, ScalaFutures}
-import org.scalatest.matchers.should.Matchers
-import org.scalatest.time.{Seconds, Span}
-import org.scalatest.{Assertion, Suite}
 import com.sksamuel.elastic4s.ElasticDsl._
 import com.sksamuel.elastic4s.requests.cluster.ClusterHealthResponse
-import com.sksamuel.elastic4s.requests.common.VersionType.ExternalGte
 import com.sksamuel.elastic4s.requests.get.GetResponse
 import com.sksamuel.elastic4s.requests.indexes.IndexResponse
 import com.sksamuel.elastic4s.requests.indexes.admin.IndexExistsResponse
 import com.sksamuel.elastic4s.requests.searches.SearchResponse
 import com.sksamuel.elastic4s.{Index, Response}
 import grizzled.slf4j.Logging
-import io.circe.{Decoder, Encoder, Json}
 import io.circe.parser.parse
-
+import io.circe.{Decoder, Encoder, Json}
+import org.scalatest.concurrent.PatienceConfiguration.Timeout
+import org.scalatest.concurrent.{Eventually, IntegrationPatience, ScalaFutures}
+import org.scalatest.matchers.should.Matchers
+import org.scalatest.time.{Seconds, Span}
+import org.scalatest.{Assertion, Suite}
 import uk.ac.wellcome.elasticsearch._
 import uk.ac.wellcome.elasticsearch.model.CanonicalId
 import uk.ac.wellcome.fixtures._
 import uk.ac.wellcome.json.JsonUtil.{fromJson, toJson}
 import uk.ac.wellcome.json.utils.JsonAssertions
-import uk.ac.wellcome.models.work.internal._
-import WorkState.Identified
+
+import scala.concurrent.ExecutionContext.Implicits.global
+import scala.concurrent.Future
 
 trait ElasticsearchFixtures
     extends Eventually
@@ -67,55 +64,10 @@ trait ElasticsearchFixtures
     response.result.numberOfNodes shouldBe 1
   }
 
-  def withLocalIndices[R](testWith: TestWith[ElasticConfig, R]): R =
-    withLocalWorksIndex { worksIndex =>
-      withLocalImagesIndex { imagesIndex =>
-        testWith(ElasticConfig(worksIndex, imagesIndex))
-      }
-    }
 
-  def withLocalWorksIndex[R](testWith: TestWith[Index, R]): R =
-    withLocalElasticsearchIndex[R](config = IndexedWorkIndexConfig) { index =>
-      testWith(index)
-    }
 
-  def withLocalSourceWorksIndex[R](testWith: TestWith[Index, R]): R =
-    withLocalElasticsearchIndex[R](config = SourceWorkIndexConfig) { index =>
-      testWith(index)
-    }
 
-  def withLocalMergedWorksIndex[R](testWith: TestWith[Index, R]): R =
-    withLocalElasticsearchIndex[R](config = MergedWorkIndexConfig) { index =>
-      testWith(index)
-    }
 
-  def withLocalIdentifiedWorksIndex[R](testWith: TestWith[Index, R]): R =
-    withLocalElasticsearchIndex[R](config = IdentifiedWorkIndexConfig) {
-      index =>
-        testWith(index)
-    }
-
-  def withLocalDenormalisedWorksIndex[R](testWith: TestWith[Index, R]): R =
-    withLocalElasticsearchIndex[R](config = DenormalisedWorkIndexConfig) {
-      index =>
-        testWith(index)
-    }
-
-  def withLocalInitialImagesIndex[R](testWith: TestWith[Index, R]): R =
-    withLocalElasticsearchIndex[R](config = InitialImageIndexConfig) { index =>
-      testWith(index)
-    }
-
-  def withLocalAugmentedImageIndex[R](testWith: TestWith[Index, R]): R =
-    withLocalElasticsearchIndex[R](config = AugmentedImageIndexConfig) {
-      index =>
-        testWith(index)
-    }
-
-  def withLocalImagesIndex[R](testWith: TestWith[Index, R]): R =
-    withLocalElasticsearchIndex[R](config = IndexedImageIndexConfig) { index =>
-      testWith(index)
-    }
 
   def withLocalElasticsearchIndex[R](
     config: IndexConfig,
@@ -158,23 +110,6 @@ trait ElasticsearchFixtures
 
       response.result.isExists shouldBe false
     }
-  }
-
-  def assertElasticsearchEventuallyHasWork[State <: WorkState](
-    index: Index,
-    works: Work[State]*)(implicit enc: Encoder[Work[State]]): Seq[Assertion] = {
-    implicit val id: CanonicalId[Work[State]] =
-      (work: Work[State]) => work.id
-    assertElasticsearchEventuallyHas(index, works: _*)
-  }
-
-  def assertElasticsearchEventuallyHasImage[State <: ImageState](
-    index: Index,
-    images: Image[State]*)(
-    implicit enc: Encoder[Image[State]]): Seq[Assertion] = {
-    implicit val id: CanonicalId[Image[State]] =
-      (image: Image[State]) => image.id
-    assertElasticsearchEventuallyHas(index, images: _*)
   }
 
   def assertElasticsearchEventuallyHas[T](index: Index, documents: T*)(
@@ -226,12 +161,7 @@ trait ElasticsearchFixtures
       hits should have size 0
     }
 
-  def assertElasticsearchNeverHasWork(index: Index,
-                                      works: Work[Identified]*): Unit = {
-    implicit val id: CanonicalId[Work[Identified]] =
-      (work: Work[Identified]) => work.state.canonicalId
-    assertElasticsearchNeverHas(index, works: _*)
-  }
+
 
   def assertElasticsearchNeverHas[T](index: Index, documents: T*)(
     implicit id: CanonicalId[T]): Unit = {
@@ -280,57 +210,9 @@ trait ElasticsearchFixtures
       }
   }
 
-  def insertIntoElasticsearch[State <: WorkState](
-    index: Index,
-    works: Work[State]*)(implicit encoder: Encoder[Work[State]]): Assertion = {
-    val result = elasticClient.execute(
-      bulk(
-        works.map { work =>
-          val jsonDoc = toJson(work).get
-          indexInto(index.name)
-            .version(work.version)
-            .versionType(ExternalGte)
-            .id(work.id)
-            .doc(jsonDoc)
-        }
-      ).refreshImmediately
-    )
 
-    // With a large number of works this can take a long time
-    // 30 seconds should be enough
-    whenReady(result, Timeout(Span(30, Seconds))) { _ =>
-      getSizeOf(index) shouldBe works.size
-    }
-  }
 
-  def insertImagesIntoElasticsearch[State <: ImageState](index: Index,
-                                                         images: Image[State]*)(
-    implicit encoder: Encoder[Image[State]]): Assertion = {
-    val result = elasticClient.execute(
-      bulk(
-        images.map { image =>
-          val jsonDoc = toJson(image).get
 
-          indexInto(index.name)
-            .version(image.version)
-            .versionType(ExternalGte)
-            .id(image.id)
-            .doc(jsonDoc)
-        }
-      ).refreshImmediately
-    )
-
-    whenReady(result) { _ =>
-      getSizeOf(index) shouldBe images.size
-    }
-  }
-
-  private def getSizeOf(index: Index): Long =
-    elasticClient
-      .execute { count(index.name) }
-      .await
-      .result
-      .count
 
   def createIndex: Index =
     Index(name = createIndexName)
